fix(game): wave-5 batch — CS_FLAGSTATUS configstring, vehicle gear-shift int compares, module-load global seeds
Team_SetFlagStatus wrote flag status to configstring 6 (CS_SCORES1)
instead of CS_FLAGSTATUS (23) — CTF flag HUD never updated and scores
were clobbered. Vehicle Update truncates prevSpeed/nextSpeed to int as
the oracle does (f32 compares flipped the gear-shift gate at fractional
speeds and desynced the Q_irand draw). GameGlobals gains a manual
Default seeding Raven's non-zero file-scope initializers:
gSiegeBeginTime=Q3_INFINITE (5s Siege warmup was skippable),
gPainHitLoc/gLastPrintedIndex=-1, saberClashEventParm=1. G_RunObject
0.7 slope compare and the holocron/shield-floor 0.1 spawn nudges run at
C's double width; Team_ForceGesture bound fixed to MAX_CLIENTS; §19
notes on st[4] zero-init and the six TIMER null guards. Adversarially
validated; full referee suite (mock + real-map duel1) green.

+++ crates/mp/game/src/g_object.rs
@@ -236,7 +236,9 @@ pub fn G_RunObject(ctx: &mut GameContext, id: EntityId) {
         // do impact physics
         if (*ent).s.pos.trType == TR_GRAVITY {
             // FIXME: only do this if no trDelta
-            if ctx.world.cvars.g_gravity.value <= 0.0f32 || tr.plane.normal[2] < 0.7f32 {
+            // `0.7` is a bare double in the oracle; promote to f64 so the
+            // round-down `<` compare matches. Source: g_object.c:196
+            if ctx.world.cvars.g_gravity.value <= 0.0f32 || (tr.plane.normal[2] as f64) < 0.7 {
                 if (*ent).flags & (FL_BOUNCE | FL_BOUNCE_HALF) != 0 {
                     if tr.fraction <= 0.0f32 {
                         crate::q_math::_VectorCopy(tr.endpos, &mut (*ent).r.currentOrigin);
